pyproject.toml
[project]
name = "code-interpreter-demo"
version = "0.1.0"
description = "Yandex Cloud AI Studio Code Interpreter demo app"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "python-multipart",
]

[dependency-groups]
dev = []
